Add support for a map query with county polygon coloring based on query results (additional variables/sliders are not yet supported - doing so will require modifications on the client and server code.
Comment-out chart/map outside clicks that are used to transition to icons to allow the explanations panel to be enlarged until explanations and a graphic display of them is ready.

Currently only the initial query is handled, subsequent 'slider' based updated queries are ignored on the server.

Numerous enhancements to address prior issues with the vue-slider control.

Move all older non-compatible example models to a new 'obsolete' dir.
Comment-out code so that the old editor is not visible from the menu (since it does not work with the new solver)
Update praise dep jar