If the map pointer returns multiple features, only return the first to mitigate map movement and give a better user experience.
If a numeric range and only displaying allowing the single value to change, only display a single dot instead of disabling the max range dot.

Wrap the Discrete values component below the Initial samples component if the width of the left panel is too small to display them both in the same row.

Display the Bar Chart and Line Graph component centered in the right panel instead of being left-aligned.