Replace mapState for MapGetters for all state access that only requires simple assignment access. Fix race condition when getting the current model name for models loaded via the UI. Do not trigger a model query modification when data is retrieved from the model editor. Completed vuex conversion for model data to the extent that the UI is fully functional and all major components access the central vuex data store rather than using property and event passing.
TBD: Relocate common code in the QueryMapResult and QueryChartResult components to the QueryGraphControls component. This includes relocating the subsequent server queries (those that use the cached server Function) to a vuex action and related debounce calling code to the QueryGraphControls component. QueryGraphControls needs to access and update the vuex data store. Currently, the results array only includes the initial HOGM queries, it needs to also include entries for each subsequent Function query. This commit includes a right-click menu to support an x-axis swap from QueryGraphControls, but additional information needs to be provided in the vuex datastore for it to function correctly and also allow the 'resultsIx change' functionality to work as expected.