Add a model to test map color assignment Add negative value support using variants of the color red:
Both red and blue color scales have 40 slots that range from white to solid red or blue.
The lowest slot for either color that is noticebly a shade of its solid color is the 3rd slot.

After determining if a number is negative, all values are rounded to 4 places.
A zero value is always assigned the zero'th slot which is white.

The lowest positive number is assigned the 3rd blue color slot and the highest is assigned the 39th slot.
All other positive values are assigned a slot relative to the highest value.

The same approach is used with the absolute value of negative numbers and the red color slots.